Akshay 13/10
changed the variables into factors. Also tried exploring on the variables but ran into the problem that there are lot of unique values in each variable. Making any graph unintelligible. We would need to talk about how to solve this and how do we show the relationship between two ordinal variables.